+++ packages/localizations/src/cs-CZ.ts
@@ -145,7 +145,7 @@ export const csCZ: LocalizationResource = {
       },
     },
     paymentMethods__label: 'Platební metody',
-    popular: 'Populární',
+    highlightedPlanBadge: 'Populární',
     pricingTable: {
       billingCycle: 'Fakturační cyklus',
       included: 'Zahrnuto',
@@ -189,6 +189,20 @@ export const csCZ: LocalizationResource = {
     },
     navbar: {
       title: 'Nastavit jednotné přihlášení (SSO)',
+    },
+    selectProviderStep: {
+      title: 'Vyberte poskytovatele',
+      subtitle: 'Vyberte poskytovatele, pro kterého budete nastavovat SSO.',
+      body: {
+        title: 'Vyberte svého poskytovatele identity',
+        description: 'Provedeme vás dále podrobným procesem nastavení.',
+      },
+      saml: {
+        groupLabel: 'SAML',
+        okta: 'Okta Workforce',
+        customSaml: 'Vlastní poskytovatel SAML',
+      },
+      warning: 'Jakmile vyberete poskytovatele, nelze ho změnit, dokud nebude konfigurace dokončena',
     },
     verifyEmailDomainStep: {
       title: 'Ověřit e-mailovou adresu',