+++ packages/localizations/src/id-ID.ts
@@ -141,7 +141,7 @@ export const idID: LocalizationResource = {
       },
     },
     paymentMethods__label: undefined,
-    popular: undefined,
+    highlightedPlanBadge: undefined,
     pricingTable: {
       billingCycle: undefined,
       included: undefined,
@@ -185,6 +185,20 @@ export const idID: LocalizationResource = {
     },
     navbar: {
       title: 'Konfigurasi Single Sign-On (SSO)',
+    },
+    selectProviderStep: {
+      title: 'Pilih penyedia',
+      subtitle: 'Pilih penyedia yang akan Anda atur untuk SSO.',
+      body: {
+        title: 'Pilih penyedia identitas Anda',
+        description: 'Kami akan memandu Anda melalui proses penyiapan terperinci berikutnya.',
+      },
+      saml: {
+        groupLabel: 'SAML',
+        okta: 'Okta Workforce',
+        customSaml: 'Penyedia SAML Khusus',
+      },
+      warning: 'Setelah penyedia dipilih, Anda tidak dapat mengubahnya lagi sampai konfigurasi selesai',
     },
     verifyEmailDomainStep: {
       title: 'Verifikasi alamat email',