+++ packages/localizations/src/sk-SK.ts
@@ -141,7 +141,7 @@ export const skSK: LocalizationResource = {
       },
     },
     paymentMethods__label: undefined,
-    popular: undefined,
+    highlightedPlanBadge: undefined,
     pricingTable: {
       billingCycle: undefined,
       included: undefined,
@@ -186,6 +186,20 @@ export const skSK: LocalizationResource = {
     },
     navbar: {
       title: 'Nastaviť jednotné prihlasovanie (SSO)',
+    },
+    selectProviderStep: {
+      title: 'Vyberte poskytovateľa',
+      subtitle: 'Vyberte poskytovateľa, pre ktorého budete nastavovať SSO.',
+      body: {
+        title: 'Vyberte si svojho poskytovateľa identity',
+        description: 'Následne vás prevedieme podrobným procesom nastavenia.',
+      },
+      saml: {
+        groupLabel: 'SAML',
+        okta: 'Okta Workforce',
+        customSaml: 'Vlastný poskytovateľ SAML',
+      },
+      warning: 'Po výbere poskytovateľa ho nemôžete zmeniť, kým sa konfigurácia neukončí',
     },
     verifyEmailDomainStep: {
       title: 'Overiť e-mailovú adresu',